Establish login page components.
